requirements.txt
agate==1.7.1
annotated-types==0.6.0
anyio==3.7.1
attrs==23.1.0
Babel==2.13.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dbt-core==1.7.0
dbt-extractor==0.5.0
dbt-postgres==1.7.0
dbt-semantic-interfaces==0.4.0
exceptiongroup==1.1.3
fastapi==0.104.1
greenlet==3.0.1
h11==0.14.0
httptools==0.6.1
idna==3.4
importlib-metadata==6.8.0
isodate==0.6.1
Jinja2==3.1.2
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
leather==0.3.4
Logbook==1.5.3
MarkupSafe==2.1.3
mashumaro==3.10
minimal-snowplow-tracker==0.0.2
more-itertools==10.1.0
msgpack==1.0.7
mysql-connector-python==8.2.0
networkx==3.2.1
packaging==23.2
parsedatetime==2.6
pathspec==0.11.2
protobuf==4.21.12
psycopg2-binary==2.9.9
pycparser==2.21
pydantic==1.10.13
pydantic_core==2.10.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
pytimeparse==1.1.8
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.6
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
sqlparse==0.4.4
starlette==0.27.0
text-unidecode==1.3
typing_extensions==4.8.0
urllib3==1.26.18
uvicorn==0.23.2
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
zipp==3.17.0
